# rest-api/openapi.yml
openapi: 3.0.0
info:
  title: Social Network & Loyalty Platform API
  version: 1.0.0
  description: |
    Combined API for social network features and loyalty program management

servers:
  - url: http://localhost:5000
    description: Local development server

paths:
  /register:
    post:
      summary: Register new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "john_doe"
                password:
                  type: string
                  format: password
                  example: "securePassword123"
              required:
                - username
                - password
      responses:
        '201':
          description: User created successfully
        '400':
          description: Invalid input

  /login:
    post:
      summary: Authenticate user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "john_doe"
                password:
                  type: string
                  format: password
                  example: "securePassword123"
              required:
                - username
                - password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Invalid credentials

  /users/{user_id}:
    get:
      summary: Get user profile
      tags: [Users]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User data
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  username:
                    type: string
                  created_at:
                    type: string
                    format: date-time
        '403':
          description: Forbidden
        '404':
          description: User not found

    put:
      summary: Update user profile
      tags: [Users]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                phone:
                  type: string
      responses:
        '200':
          description: Profile updated
        '403':
          description: Forbidden

  /posts:
    post:
      summary: Create new post
      tags: [Posts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: "My First Post"
                description:
                  type: string
                  example: "Interesting content"
                is_private:
                  type: boolean
                  default: false
                tags:
                  type: array
                  items:
                    type: string
                  example: ["tech", "news"]
              required:
                - title
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Validation error

  /posts/{post_id}:
    get:
      summary: Get post by ID
      tags: [Posts]
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Post data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '403':
          description: Access denied
        '404':
          description: Post not found

    put:
      summary: Update post
      tags: [Posts]
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                is_private:
                  type: boolean
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Post updated
        '403':
          description: Forbidden
        '404':
          description: Post not found

    delete:
      summary: Delete post
      tags: [Posts]
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Post deleted
        '403':
          description: Forbidden
        '404':
          description: Post not found

  /promos:
    post:
      summary: Create new promo code
      tags: [Promocodes]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Summer Sale"
                description:
                  type: string
                discount:
                  type: number
                  format: float
                  minimum: 0.1
                  maximum: 100
                  example: 15.5
                code:
                  type: string
                  example: "SUMMER2023"
              required:
                - name
                - discount
                - code
      responses:
        '201':
          description: Promo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promo'

  /promos/{promo_id}:
    get:
      summary: Get promo code by ID
      tags: [Promocodes]
      parameters:
        - name: promo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Promo data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promo'
        '404':
          description: Promo not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        creator_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_private:
          type: boolean
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - creator_id
        - created_at

    Promo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        creator_id:
          type: string
          format: uuid
        discount:
          type: number
          format: float
        code:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - discount
        - code
        - created_at